+++ libs/superagent/app/utils/helpers.py
