+++ libs/superagent/app/vectorstores/astra.py
@@ -9,7 +9,7 @@ from langchain.docstore.document import Document
 from langchain.embeddings.openai import OpenAIEmbeddings  # type: ignore
 from pydantic.dataclasses import dataclass
 
-from app.utils.env import get_first_non_null
+from app.utils.helpers import get_first_non_null
 from app.vectorstores.astra_client import AstraClient, QueryResponse
 
 logger = logging.getLogger(__name__)
