+++ libs/superagent/app/vectorstores/weaviate.py
@@ -9,7 +9,7 @@ from decouple import config
 from langchain.docstore.document import Document
 from langchain.embeddings.openai import OpenAIEmbeddings  # type: ignore
 from pydantic.dataclasses import dataclass
-from app.utils.env import get_first_non_null
+from app.utils.helpers import get_first_non_null
 
 logger = logging.getLogger(__name__)
